Clean up repository
This commit sets all build files I can see so far to be ignored, and untracked. This should hopefully make it easier for others to contribute, by not spamming your GitHub client with a mass of changed and untracked files. 

Also standardizes the line endings of all text files to be LF, except for .sln files which are CRLF.